Split retry handler out to separate module
