+++ pageserver/client_grpc/src/lib.rs
@@ -1,4 +1,5 @@
 mod client;
 mod pool;
+mod retry;
 
 pub use client::PageserverClient;
